app/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
elasticsearch
celery
redis
pandas
requests
jwt
python-multipart
python-dotenv==1.0.0sa